Move the HTTP endpoint to license files as well.
